moved PartialEq impl for Primitive into separate file
