+++ src/lisp/primitives/mod.rs
@@ -4,6 +4,8 @@ use std::str::FromStr;
 
 use super::{utils, LispError, LispResult, SExp};
 
+mod eq;
+
 #[derive(Clone)]
 pub enum Primitive {
     Void,
@@ -16,62 +18,6 @@ pub enum Primitive {
     Procedure(Rc<dyn Fn(SExp) -> LispResult>),
 }
 
-impl PartialEq for Primitive {
-    fn eq(&self, other: &Self) -> bool {
-        match self {
-            Primitive::Void => {
-                if let Primitive::Void = other {
-                    true
-                } else {
-                    false
-                }
-            }
-            Primitive::Undefined => {
-                if let Primitive::Undefined = other {
-                    true
-                } else {
-                    false
-                }
-            }
-            Primitive::Boolean(b1) => {
-                if let Primitive::Boolean(b2) = other {
-                    b1 == b2
-                } else {
-                    false
-                }
-            }
-            Primitive::Character(c1) => {
-                if let Primitive::Character(c2) = other {
-                    c1 == c2
-                } else {
-                    false
-                }
-            }
-            Primitive::Number(n1) => {
-                if let Primitive::Number(n2) = other {
-                    n1 == n2
-                } else {
-                    false
-                }
-            }
-            Primitive::String(s1) => {
-                if let Primitive::String(s2) = other {
-                    s1 == s2
-                } else {
-                    false
-                }
-            }
-            Primitive::Symbol(s1) => {
-                if let Primitive::Symbol(s2) = other {
-                    s1 == s2
-                } else {
-                    false
-                }
-            }
-            Primitive::Procedure(_) => false,
-        }
-    }
-}
 
 impl fmt::Debug for Primitive {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
